flask_app/requirements.txt
flask
python-dotenv
Flask-PyMongo
flask-cors